# release-21.2: backup: add simpler RESTORE span computation