--- a/pkg/ccl/backupccl/BUILD.bazel
+++ b/pkg/ccl/backupccl/BUILD.bazel
@@ -148,6 +148,7 @@ go_test(
         "restore_mid_schema_change_test.go",
         "restore_old_sequences_test.go",
         "restore_old_versions_test.go",
+        "restore_span_covering_test.go",
         "schedule_pts_chaining_test.go",
         "show_test.go",
         "split_and_scatter_processor_test.go",