--- a/pkg/ccl/backupccl/restore_job.go
+++ b/pkg/ccl/backupccl/restore_job.go
@@ -408,11 +408,23 @@ func restore(
 	// Pivot the backups, which are grouped by time, into requests for import,
 	// which are grouped by keyrange.
 	highWaterMark := job.Progress().Details.(*jobspb.Progress_Restore).Restore.HighWater
-	importSpans, _, err := makeImportSpans(dataToRestore.getSpans(), backupManifests, backupLocalityMap,
-		highWaterMark, errOnMissingRange)
-	if err != nil {
-		return emptyRowCount, errors.Wrapf(err, "making import requests for %d backups", len(backupManifests))
+
+	const useSimpleImportSpans = true
+	var importSpans []execinfrapb.RestoreSpanEntry
+	if useSimpleImportSpans {
+		if err := checkCoverage(restoreCtx, dataToRestore.getSpans(), backupManifests); err != nil {
+			return emptyRowCount, err
+		}
+		importSpans = makeSimpleImportSpans(dataToRestore.getSpans(), backupManifests, backupLocalityMap,
+			highWaterMark)
+	} else {
+		importSpans, _, err = makeImportSpans(dataToRestore.getSpans(), backupManifests, backupLocalityMap,
+			highWaterMark, errOnMissingRange)
+		if err != nil {
+			return emptyRowCount, errors.Wrapf(err, "making import requests for %d backups", len(backupManifests))
+		}
 	}
+
 	if len(importSpans) == 0 {
 		// There are no files to restore.
 		return emptyRowCount, nil